improved receipt transfer approved(created entry for if not exist in userwise table transfer to user id ) flow in ReceiptTransferService.java 180++ DPD color change added global search all term which are pending from LMS end in SearchService.java

+++ src/main/java/com/synoriq/synofin/collection/collectionservice/rest/response/globalSearchDTOs/TaskListDTOReturnResponse.java
@@ -28,12 +28,12 @@ public class TaskListDTOReturnResponse {
     String dpdBgColorKey;
 
     @JsonProperty("days_past_due")
-    Long daysPastDue;
+    int daysPastDue;
 
     @JsonProperty("address")
     String address;
 
     @JsonProperty("overdue_repayment")
-    Long overdueRepayment;
+    Double overdueRepayment;
 
 }
